Overhauled energy system to tie it directly to Mass

+++ v2/simulation/challenge.go
@@ -19,24 +19,24 @@ const (
 func PassedSurvivalCriteria(c *Creature, s *Simulation, challenge ChallengeType) bool {
 	switch challenge {
 	case LeftSurvive:
-		return c.Loc.X < float64(s.Params.GridWidth)/2
+		return c.Loc.X < s.Params.GridWidth/2
 
 	case FarLeftSurvive:
-		return c.Loc.X < float64(s.Params.GridWidth)/10
+		return c.Loc.X < s.Params.GridWidth/10
 
 	case RightSurvive:
-		return c.Loc.X > float64(s.Params.GridWidth)/2
+		return c.Loc.X > s.Params.GridWidth/2
 
 	case Groups:
 		minNeighbours := 4
 		radius := 4.0
-		isBorder := c.Loc.X < 1 || c.Loc.X >= float64(s.Params.GridWidth)-1 ||
-			c.Loc.Y < 1 || c.Loc.Y >= float64(s.Params.GridHeight)-1
+		isBorder := c.Loc.X < 1 || c.Loc.X >= s.Params.GridWidth-1 ||
+			c.Loc.Y < 1 || c.Loc.Y >= s.Params.GridHeight-1
 		if isBorder {
 			return false
 		}
-		nearEdge := c.Loc.X < 5 || c.Loc.X > float64(s.Params.GridWidth)-6 ||
-			c.Loc.Y < 5 || c.Loc.Y > float64(s.Params.GridHeight)-6
+		nearEdge := c.Loc.X < 5 || c.Loc.X > s.Params.GridWidth-6 ||
+			c.Loc.Y < 5 || c.Loc.Y > s.Params.GridHeight-6
 		if nearEdge {
 			return false
 		}
@@ -44,8 +44,8 @@ func PassedSurvivalCriteria(c *Creature, s *Simulation, challenge ChallengeType)
 		return len(neighbours)-1 >= minNeighbours // -1 to exclude self
 
 	case Center:
-		cx := float64(s.Params.GridWidth) / 2
-		cy := float64(s.Params.GridHeight) / 2
+		cx := s.Params.GridWidth / 2
+		cy := s.Params.GridHeight / 2
 		radius := 50.0
 		dx := c.Loc.X - cx
 		dy := c.Loc.Y - cy
